LPS-94205 Make XmlRpcMethodUtil unit test
