--- portal-impl/test/integration/com/liferay/portal/xmlrpc/XmlRpcMethodUtilTest.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.xmlrpc;
-
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.xmlrpc.Method;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.SyntheticBundleClassTestRule;
-import com.liferay.portal.xmlrpc.bundle.xmlrpcmethodutil.TestMethod;
-
-import org.junit.Assert;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-
-/**
- * @author Peter Fellwock
- */
-public class XmlRpcMethodUtilTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			new SyntheticBundleClassTestRule("bundle.xmlrpcmethodutil"));
-
-	@Test
-	public void testNoReturn() {
-		Method method = XmlRpcMethodUtil.getMethod(
-			TestMethod.TOKEN, TestMethod.METHOD_NAME);
-
-		Class<?> clazz = method.getClass();
-
-		Assert.assertEquals(TestMethod.class.getName(), clazz.getName());
-	}
-
-}
