--- portal-impl/test/integration/com/liferay/portal/xmlrpc/bundle/xmlrpcmethodutil/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for XmlRpcMethodUtilTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
